pyproject.toml
[project]
name = "scoring-prep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.2",
    "unbabel-comet>=2.2.7",
]
